pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Samir Mamude <mamude@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
pydantic-extra-types = "^2.9.0"



[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
coverage = "^7.6.0"
isort = "^5.13.2"
black = "^24.4.2"
pylint = "^3.2.5"
pytest-sugar = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true
source = ["src"]
